requirements.txt
# Core dependencies
PyQt6>=6.6.0
pystray>=0.19.4
plyer>=2.1.0
numpy>=1.21.0

# Audio processing
sounddevice>=0.4.6
PyAudio>=0.2.11

# AI and ML - Speech Recognition
openai-whisper>=20231117
torch>=2.1.0
transformers>=4.35.0

# Alternative lightweight options
# faster-whisper>=0.10.0  # Lighter alternative to openai-whisper
# vosk>=0.3.45           # Offline speech recognition

# Screen capture and image processing
mss>=9.0.1
Pillow>=10.1.0

# Data and scheduling
schedule>=1.2.0
# sqlite3 is built-in

# Utilities and OS integration
python-dotenv>=1.0.0
psutil>=5.9.6
requests>=2.31.0

# Windows-specific packages (optional)
# pywin32>=306          # For Windows API access
# winshell>=0.6         # For Windows shortcuts
# pycaw>=20230407       # For Windows audio control

# Future: LLM integration
# huggingface-hub>=0.17.0
# accelerate>=0.24.0